Create paper_wallet_seed_generator
This program is used to pic the first 23 seed words for a Paper Wallet.  It should not be used on a computer connected to the Internet. You should really only use it on TAILS.

Be very careful.

Using this for a live wallet on the Internet is an obvious security hole. 

It is not a good idea to create bitcoin wallet seeds on machines connected to the Internet.